[Tue-15.12-16:52] [DEV] norme OK, compil helper OK, smoke tests OK [NEXT] philo_two

+++ philo_one/srcs/utils/status_printer.c
@@ -6,7 +6,7 @@
 /*   By: bvalette <bvalette@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/09 12:08:58 by bvalette          #+#    #+#             */
-/*   Updated: 2020/12/15 15:46:41 by bvalette         ###   ########.fr       */
+/*   Updated: 2020/12/15 16:47:41 by bvalette         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,8 +19,8 @@ void	put_regular_status(t_data *data, const int philo_id,
 	if (data->first_death_report == false)
 	{
 		ft_putnbr(STDOUT_FILENO, data->philo_state_time_stamp[philo_id]);
-		ft_putnbr(STDOUT_FILENO, philo_id);
-		ft_put_message_fd(STDOUT_FILENO, message_len, message);
+		ft_putnbr(STDOUT_FILENO, (unsigned long)philo_id);
+		ft_put_message_fd(STDOUT_FILENO, (size_t)message_len, message);
 	}
 	pthread_mutex_unlock(&data->mutex_stdout);
 }
@@ -36,7 +36,7 @@ void	put_death_status(t_data *data, const int philo_id)
 		{
 			already_reported_flag = true;
 			ft_putnbr(STDOUT_FILENO, data->first_death_report_timestamp);
-			ft_putnbr(STDOUT_FILENO, philo_id);
+			ft_putnbr(STDOUT_FILENO, (unsigned long)philo_id);
 			ft_put_message_fd(STDOUT_FILENO, LEN_IS_DEAD, MESSAGE_IS_DEAD);
 		}
 		pthread_mutex_unlock(&data->mutex_stdout);
